refactor: refine dynamic string APIs and clean up code
- Rename yio_sprint to yio_snstream for naming consistency.
- Rename yio_append_f to yio_append (with data) and adjust no-data macro to yio_append_s.
- Remove unnecessary casts and brackets across multiple formatting modules.
- Simplify pfmt_parse test using a static zero-initialized structure.
- Clarify YYIO_LIMIT_MAX as a safety cap for 8-bit formatting fields.
- Fix logic in yio_vasprintf and yio_vappend to correctly handle storage reuse and offsets.

+++ src/yio/private/yio_float_strfrom_custom.c
@@ -227,7 +227,7 @@ int YYIO_float_astrfrom_custom$1(YYIO_string *v, int precision0, char spec0, TYP
 	// The printed exponent is one less, cause of the initial digit!
 	int exponent = 0;
 
-	precision = (int)yio_precision_get_default(precision0, spec0lower == 'a' ? a_max_precision : 6);
+	precision = yio_precision_get_default(precision0, spec0lower == 'a' ? a_max_precision : 6);
 
 	// Rounding with 'e' specifier is shared with 'g' in case it get's chosen
 	int exponent10 = 0;
